add Makefile.PL
I have no idea what I'm doing but I require the change from the pull request for our CI
and this uses cpanm which requires a Makefile.PL so I searched for a solution and found this:
http://blogs.perl.org/users/piotr_roszatycki/2014/01/i-hate-distzilla-and-missing-makefilepl.html
and because of the extra work I have to do to use the fixed module I agree with the author.
this solution appears to be working, so: yay!