Passthrough additional params to is_holidays functions
In the Netherlands (and Aruba) some holidays are not for all people, but just
for government employees. Most notable being liberation day. I want to pass the
params to the is_holiday and is_holiday_dt methods to the actual implemenation
so I can get the params hash for `gov => 1, lang => nl` in both functions.

Signed-off-by: Wesley Schwengle <waterkip@cpan.org>